Added more test coverage
