+++ library/a-star/.gitignore
@@ -1,2 +1,4 @@
 .idea
 out
+.project
+.classpath
