+++ library/a-star/bin/.gitignore
@@ -1,0 +1,2 @@
+/main/
+/test/
