+++ library/a-star/build.gradle
@@ -15,5 +15,7 @@ repositories {
 dependencies{
 	testImplementation 'junit:junit:4.13'
 	implementation group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
+	implementation group: 'org.imgscalr', name: 'imgscalr-lib', version: '4.2'
+	
 }
