+++ tools/metadata-conversion-tool/.gitignore
@@ -1,0 +1,4 @@
+.project
+.classpath
+/.idea/
+build/
