+++ tools/metadata-conversion-tool/bin/.gitignore
@@ -1,0 +1,4 @@
+/main/
+.project
+.classpath
+.idea/
