+++ tools/metadata-conversion-tool/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'metadata-conversion-tool'
+
